Fixed some issues added cart view

+++ src/Modules/Sales/ECommerce.Modules.Sales.Application/Orders/Commands/Handlers/DeleteOrderItemFromOrderHandler.cs
@@ -1,4 +1,5 @@
 ﻿using ECommerce.Modules.Sales.Application.Orders.Exceptions;
+using ECommerce.Modules.Sales.Application.Orders.Policies;
 using ECommerce.Modules.Sales.Domain.Orders.Repositories;
 using ECommerce.Shared.Abstractions.Commands;
 using System;
@@ -13,11 +14,13 @@ namespace ECommerce.Modules.Sales.Application.Orders.Commands.Handlers
     {
         private readonly IOrderRepository _orderRepository;
         private readonly IOrderItemRepository _orderItemRepository;
+        private readonly IOrderPositionModificationPolicy _orderPositionModificationPolicy;
 
-        public DeleteOrderItemFromOrderHandler(IOrderRepository orderRepository, IOrderItemRepository orderItemRepository)
+        public DeleteOrderItemFromOrderHandler(IOrderRepository orderRepository, IOrderItemRepository orderItemRepository, IOrderPositionModificationPolicy orderPositionModificationPolicy)
         {
             _orderRepository = orderRepository;
             _orderItemRepository = orderItemRepository;
+            _orderPositionModificationPolicy = orderPositionModificationPolicy;
         }
 
         public async Task HandleAsync(DeleteOrderItemFromOrder command)
@@ -30,9 +33,23 @@ namespace ECommerce.Modules.Sales.Application.Orders.Commands.Handlers
             }
 
             var order = await _orderRepository.GetDetailsAsync(command.OrderId);
+
+            if (order is null)
+            {
+                throw new OrderNotFoundException(command.OrderId);
+            }
+
+            var canDeletePosition = await _orderPositionModificationPolicy.CanDeleteAsync(order);
+            if (!canDeletePosition)
+            {
+                throw new PositionFromOrderCannotBeDeletedException(order.Id, command.OrderItemId);
+            }
+
             order.DeleteOrderItem(orderItem);
+            order.RefreshCost();
 
             await _orderRepository.UpdateAsync(order);
+            await _orderItemRepository.DeleteAsync(orderItem);
         }
     }
 }
